# charts: Add mount for wasm-cache