@mqasimsarfraz 2025-07-17
> Why is this cache needed and what are the files cached? 

The cache helps reduce the startup time for a gadget because of `Runtime.CompileModule`, especially since we compile module both for `GetGadgetInfo` and `RunGadget`. Only the compiled module is cached, you can also check the content of `/var/run/ig/wasm-cache/wazero-v1.9.0-amd64-linux` to see what is being cache for each gadget run for `ig`. 


> Machine code translated from WASM files in gadgets by wazero?

Yes, following are couple sources:

>  CompilationCache reduces time spent compiling (Runtime.CompileModule) the same wasm module.

https://pkg.go.dev/github.com/tetratelabs/wazero#CompilationCache

> Cache allows the compiler engine to skip compilation of wasm to machine code where doing so is redundant for the same wasm binary and version of wazero.

https://pkg.go.dev/wa-lang.org/wazero/internal/compilationcache?#pkg-types