@mqasimsarfraz 2025-08-06
> I'm bumping ig on Gentoo and have noticed the new warning, simply by running ig.
> 
> ```
> $ ig
> WARN[0000] failed to setup wasm compilation cache, skipping cache: create directory /var/run/ig/wasm-cache: mkdir /var/run/ig: permission denied 
> ```

Oops, this is annoying. 

> Or maybe you're always supposed to run ig as root? 

Yes, `ig` should always run as root and agreed we should change the logic to skip the warning when ran as non-root user. Let me have a look into it! :)